+++ basic_autonomy_ros2/src/basic_autonomy.cpp
@@ -817,9 +817,9 @@ namespace basic_autonomy
                 size_t left_points_size = basic_points.size() - resized_basic_points.size();
                 RCLCPP_DEBUG_STREAM(rclcpp::get_logger(BASIC_AUTONOMY_LOGGER), "Number of left out basic_points size: " << left_points_size);
 
-                int percent_points_lost = 100 * left_points_size/basic_points.size();
+                double percent_points_lost = 100 * left_points_size/basic_points.size();
 
-                if (percent_points_lost > 50)
+                if (percent_points_lost > 50.0)
                 {
                     RCLCPP_WARN_STREAM(rclcpp::get_logger(BASIC_AUTONOMY_LOGGER), "More than half of basic points are ignored for spline fitting");
                 }
